fix(typecheck): repo-wide type-error cleanup — all green except integration-hub (WIP)
Root cause: the `baseUrl` TS5101 deprecation was a *blocking* config error that
halted per-project typecheck before file checking, masking real drift; only 7 of
~40 projects even had a CI typecheck target, so it rotted invisibly.

Config
- centralize `ignoreDeprecations: "6.0"` in tsconfig.base (kills 21 baseUrl errors)
- add @types/node to ts-sdk-api-key (uses node:crypto/Buffer); drop unused node
  types from ts-sdk-india / ts-sdk-sequence

Pulse (all our-side wiring; nothing broken on IQ-Line's end)
- fix stale vendored tsconfig paths (../ui -> ../pulse-ui, etc.) from the rename
- back-button: leftover react-router API -> TanStack idiom (router.history.back)
- import.meta.env: self-contained ambient decl. All 4 pulse pkgs -> 0 errors

user-management (28+13 -> 0; 290+27 tests green)
- RecoveryTier typing (no casts), missing must_change_password select (real bug),
  duplicate-identifier dedup, test mock/contract fixes (removed an `as never`)

clinical/billing (billing/pharmacy/inventory -> 0; registration -> reports-only)
- real bug: pharmacy actorIdFromRequest read wrong Principal field (createdBy
  was always null); stray fastify.d.ts augmentation corrupting shared identity types

pdf-client/web: node types + scan-share prefill typing (web in-scope -> 0)

python (all 7 projects): opd pyright 37 -> 0 (only configured checker), ruff 7/7,
real Decimal-vs-float ORM annotation fix; 512 tests green

integration-hub fix is in progress (uncommitted, on disk) — to finish next session.
Also updates CLAUDE.md: tsc is fine to run one-shot on the current WSL2 setup.

+++ modules/opd/src/opd/models/prescription/children.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import UTC, date, datetime
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from sqlalchemy import (
@@ -65,16 +66,16 @@ class PrescriptionLegacyVitalsModel(TimestampMixin, TenantPrimaryKeyMixin, Base)
     prescription_id: Mapped[uuid.UUID] = mapped_column(
         Uuid(as_uuid=True), primary_key=True, nullable=False
     )
-    height_cm: Mapped[float | None] = mapped_column(Numeric(6, 2), nullable=True)
-    weight_kg: Mapped[float | None] = mapped_column(Numeric(6, 2), nullable=True)
-    bmi: Mapped[float | None] = mapped_column(Numeric(5, 2), nullable=True)
-    temperature_c: Mapped[float | None] = mapped_column(Numeric(4, 1), nullable=True)
+    height_cm: Mapped[Decimal | None] = mapped_column(Numeric(6, 2), nullable=True)
+    weight_kg: Mapped[Decimal | None] = mapped_column(Numeric(6, 2), nullable=True)
+    bmi: Mapped[Decimal | None] = mapped_column(Numeric(5, 2), nullable=True)
+    temperature_c: Mapped[Decimal | None] = mapped_column(Numeric(4, 1), nullable=True)
     pulse_bpm: Mapped[int | None] = mapped_column(SmallInteger, nullable=True)
     bp_systolic: Mapped[int | None] = mapped_column(SmallInteger, nullable=True)
     bp_diastolic: Mapped[int | None] = mapped_column(SmallInteger, nullable=True)
     respiratory_rate: Mapped[int | None] = mapped_column(SmallInteger, nullable=True)
     spo2_percent: Mapped[int | None] = mapped_column(SmallInteger, nullable=True)
-    blood_sugar_mg_dl: Mapped[float | None] = mapped_column(Numeric(6, 1), nullable=True)
+    blood_sugar_mg_dl: Mapped[Decimal | None] = mapped_column(Numeric(6, 1), nullable=True)
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
 
     prescription: Mapped[PrescriptionModel] = relationship(back_populates="legacy_vitals")
@@ -255,7 +256,7 @@ class PrescriptionMedicineModel(TimestampMixin, LineItemMixin, Base):
     dosage: Mapped[str | None] = mapped_column(String(256), nullable=True)
     duration: Mapped[str | None] = mapped_column(String(128), nullable=True)
     frequency: Mapped[str | None] = mapped_column(String(128), nullable=True)
-    quantity: Mapped[float | None] = mapped_column(Numeric(10, 2), nullable=True)
+    quantity: Mapped[Decimal | None] = mapped_column(Numeric(10, 2), nullable=True)
     route: Mapped[str | None] = mapped_column(String(64), nullable=True)
     method: Mapped[str | None] = mapped_column(String(128), nullable=True)
     status: Mapped[str | None] = mapped_column(String(32), nullable=True)
@@ -279,7 +280,7 @@ class PrescriptionMedicineSubstitutionModel(TimestampMixin, TenantPrimaryKeyMixi
     issued_medicine_id: Mapped[uuid.UUID | None] = mapped_column(Uuid(as_uuid=True), nullable=True)
     issued_name: Mapped[str] = mapped_column(String(512), nullable=False)
     item_code: Mapped[str | None] = mapped_column(String(64), nullable=True)
-    quantity: Mapped[float | None] = mapped_column(Numeric(10, 2), nullable=True)
+    quantity: Mapped[Decimal | None] = mapped_column(Numeric(10, 2), nullable=True)
     form: Mapped[str | None] = mapped_column(String(128), nullable=True)
     volume: Mapped[str | None] = mapped_column(String(64), nullable=True)
     category: Mapped[str | None] = mapped_column(String(128), nullable=True)
